fix: force server side rendering
