+++ frontend/src/app/page.tsx
@@ -1,17 +1,16 @@
-"use client";
-
-import { Heading } from "@/components/Heading";
-
-export const dynamic = "force-dynamic";
-
-const Content = (): JSX.Element => {
-  return <Heading />;
-};
+import { Suspense } from "react";
+import { Loading } from "@/components/Loading";
+import { StatusFetcher } from "@/components/StatusFetcher";
 
 export default function Home() {
   return (
     <div className="w-full items-center">
-      <Content />
+      <div className="flex items-center gap-4 text-4xl font-extrabold leading-none tracking-tight text-gray-900 md:text-5xl lg:text-6xl ">
+        <h1>Server is</h1>
+        <Suspense fallback={<Loading />}>
+          <StatusFetcher />
+        </Suspense>
+      </div>
     </div>
   );
 }
